consolidate step code requirement table
